Debugger: Sanitize Datatypes, part 1: Breakpoints.
Make address a quint64, linnumber and ignoreCounts int.

+++ src/plugins/debugger/qml/qmlengine.cpp
@@ -439,7 +439,7 @@ void QmlEngine::attemptBreakpointSynchronization()
         QString processedFilename = data->fileName;
         if (isShadowBuildProject())
             processedFilename = toShadowBuildFilename(data->fileName);
-        breakList << qMakePair(processedFilename, data->lineNumber.toInt());
+        breakList << qMakePair(processedFilename, data->lineNumber);
     }
 
     {
